+++ docs/CHANGELOG.md
@@ -39,6 +39,7 @@
     - `trx.input.is_key_held`
     - `trx.input.is_key_pressed`
     - `trx.input.is_key_known`
+    - `trx.input.is_reserved`
 - Added the ability for a script to read the controller itself, sticks and triggers included
     - `trx.events.on_button_down`
     - `trx.events.on_button_up`